requirements.txt
groq
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-groq
datasets
spacy
streamlit
SPARQLWrapper
nltk
rouge-score
plotly
pandas
numpy
transformers
torch
python-dotenv